op-challenger: Combine FaultDisputeGameCaller and Loader into a single wrapper for the fault dispute game contract.
