+++ op-challenger/fault/contracts/fault_dispute_game.go
@@ -1,16 +1,19 @@
-package fault
+package contracts
 
 import (
 	"context"
 	"math/big"
 
+	"github.com/ethereum-optimism/optimism/op-bindings/bindings"
 	"github.com/ethereum-optimism/optimism/op-challenger/fault/types"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 )
 
 // MinimalFaultDisputeGameCaller is a minimal interface around [bindings.FaultDisputeGameCaller].
 // This needs to be updated if the [bindings.FaultDisputeGameCaller] interface changes.
 type MinimalFaultDisputeGameCaller interface {
+	Status(opts *bind.CallOpts) (uint8, error)
 	ClaimData(opts *bind.CallOpts, arg0 *big.Int) (struct {
 		ParentIndex uint32
 		Countered   bool
@@ -23,32 +26,32 @@ type MinimalFaultDisputeGameCaller interface {
 	ABSOLUTEPRESTATE(opts *bind.CallOpts) ([32]byte, error)
 }
 
-// Loader is a minimal interface for loading onchain [Claim] data.
-type Loader interface {
-	FetchClaims(ctx context.Context) ([]types.Claim, error)
-	FetchGameDepth(ctx context.Context) (uint64, error)
-	FetchAbsolutePrestateHash(ctx context.Context) ([]byte, error)
-}
-
-// loader pulls in fault dispute game claim data periodically and over subscriptions.
-type loader struct {
+// FaultDisputeGame provides an anti-corruption layer between the on-chain contracts and op-challenger bounded contexts.
+// It improves the decoupling between these two contexts to make it easier to adapt to any future API changes.
+// Additionally, it provides a simpler API for the rest of op-challenger to use with better typing of values.
+type FaultDisputeGame struct {
 	caller MinimalFaultDisputeGameCaller
 }
 
-// NewLoader creates a new [loader].
-func NewLoader(caller MinimalFaultDisputeGameCaller) *loader {
-	return &loader{
-		caller: caller,
+func NewFaultDisputeGame(caller MinimalFaultDisputeGameCaller) *FaultDisputeGame {
+	return &FaultDisputeGame{caller: caller}
+}
+
+func NewFaultDisputeGameAtAddress(addr common.Address, client bind.ContractCaller) (*FaultDisputeGame, error) {
+	caller, err := bindings.NewFaultDisputeGameCaller(addr, client)
+	if err != nil {
+		return nil, err
 	}
+	return NewFaultDisputeGame(caller), nil
 }
 
 // FetchGameDepth fetches the game depth from the fault dispute game.
-func (l *loader) FetchGameDepth(ctx context.Context) (uint64, error) {
+func (g *FaultDisputeGame) FetchGameDepth(ctx context.Context) (uint64, error) {
 	callOpts := bind.CallOpts{
 		Context: ctx,
 	}
 
-	gameDepth, err := l.caller.MAXGAMEDEPTH(&callOpts)
+	gameDepth, err := g.caller.MAXGAMEDEPTH(&callOpts)
 	if err != nil {
 		return 0, err
 	}
@@ -57,12 +60,12 @@ func (l *loader) FetchGameDepth(ctx context.Context) (uint64, error) {
 }
 
 // fetchClaim fetches a single [Claim] with a hydrated parent.
-func (l *loader) fetchClaim(ctx context.Context, arrIndex uint64) (types.Claim, error) {
+func (g *FaultDisputeGame) fetchClaim(ctx context.Context, arrIndex uint64) (types.Claim, error) {
 	callOpts := bind.CallOpts{
 		Context: ctx,
 	}
 
-	fetchedClaim, err := l.caller.ClaimData(&callOpts, new(big.Int).SetUint64(arrIndex))
+	fetchedClaim, err := g.caller.ClaimData(&callOpts, new(big.Int).SetUint64(arrIndex))
 	if err != nil {
 		return types.Claim{}, err
 	}
@@ -80,7 +83,7 @@ func (l *loader) fetchClaim(ctx context.Context, arrIndex uint64) (types.Claim, 
 
 	if !claim.IsRootPosition() {
 		parentIndex := uint64(fetchedClaim.ParentIndex)
-		parentClaim, err := l.caller.ClaimData(&callOpts, new(big.Int).SetUint64(parentIndex))
+		parentClaim, err := g.caller.ClaimData(&callOpts, new(big.Int).SetUint64(parentIndex))
 		if err != nil {
 			return types.Claim{}, err
 		}
@@ -94,9 +97,9 @@ func (l *loader) fetchClaim(ctx context.Context, arrIndex uint64) (types.Claim, 
 }
 
 // FetchClaims fetches all claims from the fault dispute game.
-func (l *loader) FetchClaims(ctx context.Context) ([]types.Claim, error) {
+func (g *FaultDisputeGame) FetchClaims(ctx context.Context) ([]types.Claim, error) {
 	// Get the current claim count.
-	claimCount, err := l.caller.ClaimDataLen(&bind.CallOpts{
+	claimCount, err := g.caller.ClaimDataLen(&bind.CallOpts{
 		Context: ctx,
 	})
 	if err != nil {
@@ -106,7 +109,7 @@ func (l *loader) FetchClaims(ctx context.Context) ([]types.Claim, error) {
 	// Fetch each claim and build a list.
 	claimList := make([]types.Claim, claimCount.Uint64())
 	for i := uint64(0); i < claimCount.Uint64(); i++ {
-		claim, err := l.fetchClaim(ctx, i)
+		claim, err := g.fetchClaim(ctx, i)
 		if err != nil {
 			return nil, err
 		}
@@ -117,16 +120,29 @@ func (l *loader) FetchClaims(ctx context.Context) ([]types.Claim, error) {
 }
 
 // FetchAbsolutePrestateHash fetches the hashed absolute prestate from the fault dispute game.
-func (l *loader) FetchAbsolutePrestateHash(ctx context.Context) ([]byte, error) {
+func (g *FaultDisputeGame) FetchAbsolutePrestateHash(ctx context.Context) (common.Hash, error) {
 	callOpts := bind.CallOpts{
 		Context: ctx,
 	}
 
-	absolutePrestate, err := l.caller.ABSOLUTEPRESTATE(&callOpts)
+	absolutePrestate, err := g.caller.ABSOLUTEPRESTATE(&callOpts)
 	if err != nil {
-		return nil, err
+		return common.Hash{}, err
 	}
-	returnValue := absolutePrestate[:]
+	return absolutePrestate, nil
+}
 
-	return returnValue, nil
+// GetGameStatus returns the current game status.
+func (g *FaultDisputeGame) GetGameStatus(ctx context.Context) (types.GameStatus, error) {
+	status, err := g.caller.Status(&bind.CallOpts{Context: ctx})
+	return types.GameStatus(status), err
+}
+
+// GetClaimCount returns the number of claims in the game.
+func (g *FaultDisputeGame) GetClaimCount(ctx context.Context) (uint64, error) {
+	count, err := g.caller.ClaimDataLen(&bind.CallOpts{Context: ctx})
+	if err != nil {
+		return 0, err
+	}
+	return count.Uint64(), nil
 }
